Extract trckvetoDep the PAT way, i.e. from isodeposit candenergy, as is done for the ecal and hcal, rather than from the inherited muon object. Btw it kinda bothers me that it's trckvetoDep when we have trackIso, but I'm going to resist changing these names here and in the MuonMaker.
In the same vein, get iso values the PAT way, i.e. userIsolation(pat::blah),
rather than from the inherited muon as the pat::[track|ecal|hcal]Iso methods
do.

It's all rather silly, they are exactly the same. We could just as well have
left the iso values as they were and instead extracted the vetoDeps from the
reco::Muon.  However, since this is PATMuonMaker we'll do things the PAT way
and ensure consistency between our iso and vetoDep values.

Renamed mus_pat_vetoDep to mus_pat_calovetoDep so that I don't have to check
the code to see that it does not include the tracker. Also we have caloIso
and not just iso so it kinda makes sense. I checked that this is not used by
anyone anywhere so probably we can just drop it altogether as it is trivial to
recalculate. mus_pat_caloIso is also trivial to recalculate...